CORRECT: put, delete 된 후 label 페이지 재 랜더링
- 수정된 사항이 반영되고 바로 페이지가 재 렌더링이 됩니다.
- post는 아직 반영이 안되서 추후 fix 예정입니다.
